requirements.txt
# Requirements:
Flask==2.0.2
gunicorn==20.1.0
python-dateutil==2.8.2
pytz==2021.3

# Dependencies:
click==8.0.3
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
six==1.16.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0